+++ scm-core/src/main/java/sonia/scm/repository/api/BrowseCommandBuilder.java
@@ -300,6 +300,35 @@ public final class BrowseCommandBuilder
     return this;
   }
 
+  /**
+   * Limit the number of result files to <code>limit</code> entries. By default this is set to
+   * {@value BrowseCommandRequest#DEFAULT_REQUEST_LIMIT}. Be aware that this parameter can have
+   * severe performance implications. Reading a repository with thousands of files in one folder
+   * can generate a huge load for a longer time.
+   *
+   * @param limit The maximal number of files this request shall return (directories are <b>not</b> counted).
+   *
+   * @since 2.0.0
+   */
+  public BrowseCommandBuilder setLimit(int limit) {
+    request.setLimit(limit);
+    return this;
+  }
+
+  /**
+   * Proceed the list from the given number on (zero based).
+   *
+   * @param offset The number of the file, the result should start with (zero based).
+   *               All preceding files will be omitted. Directories are <b>not</b>
+   *               counted. Therefore directories are only listed in results without
+   *               offset.
+   * @since 2.0.0
+   */
+  public BrowseCommandBuilder setOffset(int offset) {
+    request.setOffset(offset);
+    return this;
+  }
+
   private void updateCache(BrowserResult updatedResult) {
     if (!disableCache) {
       CacheKey key = new CacheKey(repository, request);