+++ scm-plugins/scm-hg-plugin/src/main/java/sonia/scm/repository/spi/HgBrowseCommand.java
@@ -101,6 +101,9 @@ public class HgBrowseCommand extends AbstractCommand implements BrowseCommand
       cmd.disableSubRepositoryDetection();
     }
 
+    cmd.setLimit(request.getLimit());
+    cmd.setOffset(request.getOffset());
+
     FileObject file = cmd.execute();
     return new BrowserResult(c == null? "tip": c.getNode(), revision, file);
   }